+++ CHANGELOG.md
@@ -2,7 +2,15 @@
 
 ## [Version - 1]
 
-## Added [v1.4.0] - 2020-1-7
+## Fixed [v1.5.2] - 2020-1-7
+- Modal bounce in delivery method
+
+## Added [v1.5.1] - 2020-1-7
+- Fixed add product to cart on text change
+- Fixed bug ket prop
+- Auto Fill zero in Input telephon screen
+
+## Added [v1.5.0] - 2020-1-7
 - Can choose a delivery method
 
 ## Fixed [v1.4.1] - 2020-1-5